Fire property change function implemented with java doc
